webpack2 and typescript integration
