+++ package.json
@@ -1,13 +1,14 @@
 {
   "dependencies": {
+    "@types/react": "^15.0.6",
     "autoprefixer": "^6.5.1",
+    "axios": "^0.15.3",
     "classnames": "^2.2.5",
     "express": "^4.14.0",
     "huntjs": "^2.0.1",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^4.16.6",
     "moment": "^2.15.2",
-    "qs": "^6.3.0",
     "react": "^15.3.2",
     "react-dom": "^15.3.2",
     "react-gateway": "^2.3.0",
@@ -16,9 +17,11 @@
     "react-router": "^3.0.0",
     "react-select": "^1.0.0-rc.2",
     "request": "^2.78.0",
+    "ts-loader": "^2.0.0",
     "webpack-dev-middleware": "^1.8.4"
   },
   "devDependencies": {
+    "@types/axios": "^0.9.35",
     "babel-core": "^6.18.2",
     "babel-loader": "^6.2.7",
     "babel-preset-react-app": "^1.0.0",
@@ -26,15 +29,15 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "less": "^2.7.1",
     "less-loader": "^2.2.3",
-    "nodemon": "^1.11.0",
     "postcss": "^5.2.5",
     "postcss-loader": "^1.1.0",
     "style-loader": "^0.13.1",
+    "typescript": "^2.1.5",
     "webpack": "^1.13.3",
     "webpack-dev-server": "^1.16.2"
   },
   "scripts": {
-    "start": "NODE_ENV=development node_modules/.bin/nodemon src/server.js",
+    "start": "NODE_ENV=development node src/server.js",
     "build": "NODE_ENV=production node_modules/.bin/webpack"
   }
 }
