+++ webpack.config.js
@@ -2,11 +2,14 @@ const path = require("path")
 const ExtractTextPlugin = require("extract-text-webpack-plugin")
 
 module.exports = {
-  entry: path.join(__dirname, 'src', 'client.js'),
+  entry: './src/client',
   output: {
     filename: 'bundle.js',
     path: path.join(__dirname, '/build'),
     publicPath: 'build'
+  },
+  resolve: {
+    extensions: ['', '.js', '.ts', '.tsx']
   },
   module: {
     loaders: [
@@ -15,6 +18,7 @@ module.exports = {
         loader: 'babel-loader',
         exclude: /node_modules/
       },
+      { test: /\.ts(x?)$/, loader: 'babel-loader!ts-loader' },
       {
         test: /\.css$/,
         loader: ExtractTextPlugin.extract('style-loader', 'css-loader!postcss-loader')
